style: followController 생성
